Update requirements.txt
Someone put a very stale version (0.3.0 - current is 1.2.0) of the read the docs theme in the repo. Doug has a newer version installed on his system, which gets used when he builds the docs. The github action doesn't install the theme, so it must rely on the old one. Recommendation: add sphinx-rtd-theme to the requirements.txt and delete the stale theme from the _themes directory. This will likely fix it.